requirements.txt
appdirs==1.4.0
Django==1.8
django-annoying==0.10.3
django-crispy-forms==1.6.1
django-selectize==1.2
packaging==16.8
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
